requirements.txt
chainlit
requests
openai
ollama
streamlit
docker
python-dotenv
uv
mcp-obsidian